requirements.txt
streamlit==1.38.0
openai>=1.0.0
python-dotenv
